✏️ Move ClickMeButton in its own file.
