+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/CallCardScreenshots/CallCardPreviewTest_Dark_d19fbf1f_0.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:857d1a8e1b2e09c15703caacad2551569fb912ff860612e0e69335f6cfc72005
-size 3944
+oid sha256:1470ad60f384c780392ec823652a5b7beecbc4e8a47870e68c6f9fe1548be60b
+size 3915
