+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/CallCardScreenshots/CallCardPreviewTest_Landscape Mode_67531073_0.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:8063a9ee6d2f56a35d4cdb66e57b183fd7113731fb2f6ca2981d9e397932ba91
-size 1851
+oid sha256:f748315719094510d9c6ef71db459c2ae0cb06e4b0726d52fc5400b9a9449a45
+size 1832
