+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/CallCardScreenshots/CallCardPreviewTest_Large Font Size_8b5d39f1_0.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f937712f62e5180c7dd280237f8e58219e39f063bc94b65758a2a3e97193d6bc
-size 4652
+oid sha256:9e5a918ab8e6a706ceb0f74fe843f53633800a8c13f293ac74d5981173b6aa14
+size 4635
