+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/CallCardScreenshots/CallCardPreviewTest_Portrait Mode_3f51852e_0.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:511637e08ed303d6beafc21e9e38f694353470a78ef5d5a53a035c72d5bf8a39
-size 3817
+oid sha256:0f473ad25dcced9774d21f41788c19e0efd6877d037467521c7dc935950ba444
+size 3786
