+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ClickMeButtonScreenshots/ButtonPreviewTest_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:bcd0226fc84136ee0ed1d215cacb96d7645c3ba23c37bdd452272d44023f7c5a
+size 4427
