+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ClickMeButtonScreenshots/ButtonPreviewTest_3ec4aff3_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d5a20eece52ed563b586425df41c088a650f997870e53d91622ed4b033c50ed7
+size 5264
