+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ClickMeButtonScreenshots/ButtonPreviewTest_Dark_d19fbf1f_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:117ec21ce668779fdec439e404ee705e590f537a028b2c00ffa5ea5253cb3621
+size 4275
