+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ClickMeButtonScreenshots/ButtonPreviewTest_Landscape Mode_67531073_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:78c425d88f4f03ee02d106c6a9e81b0bfed722866d4540f71be89ba9555e9328
+size 1996
