+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ClickMeButtonScreenshots/ButtonPreviewTest_Large Font Size_8b5d39f1_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6677d26116042d74297000ef5f87aac13e253f4defac22d2582f419557650de5
+size 5982
