--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_328e5253_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f91336ca672a79b42eb7a74137b30982981a7875cb1188cc04954c87ca8848c7
-size 105871
