--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_3ec4aff3_328e5253_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0768ff83e3034fb8c30e12cb639e018fa6b0a0e2b53f098eb9f82054fd6950f2
-size 105111
