+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_3ec4aff3_4310119c_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6480b3bb55183e407c61b136ee6a83fb5c0c24833d7e96cda8d66180aad5909b
+size 104424
