--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Dark_d19fbf1f_328e5253_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4f85bb5c2fc42ea7496c64267b9ea97f88e255dedc0d208c0e3ecef6aff8cab1
-size 107063
