+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Dark_d19fbf1f_4310119c_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:822b6e70233339214162153565f0dadb33ca19f6bd22a7154e9210614c1f47ee
+size 106611
