--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Landscape Mode_67531073_328e5253_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:5cdc7f99bc2eba2993e245f0c43d93be24d4b38e4a9b5246b5293c7e5f02d5b3
-size 34109
