+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Landscape Mode_67531073_4310119c_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:97d058052d0b41b879e5fa3af44f02a25940dd93ddcc65f21902c95bdb4117b6
+size 33787
