--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Large Font Size_8b5d39f1_328e5253_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:538670d5c5c286084254a880d8dad15e4d02664894e64b43129d8f9e6332f028
-size 138852
