+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Large Font Size_8b5d39f1_4310119c_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:abc9aa2de06a981d3a03b4778cfdbd8fdbc47e468a30ac8b8979a4fa7f95481b
+size 138270
