+++ app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/DirectoryScreenScreenshots/CallCardPreviewTest_Light_b29dc7a7_4310119c_0.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2471b0c4133c870da6953833bd48eeea86b3a945f0a03af18a8fac650dadf2b4
+size 105131
