--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ExamplePreviewsScreenshots/ButtonPreviewTest_3ec4aff3_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:782ba4c6277c83648623e2fb31709ab46f5dc4621982a7e3feb85a6ceb006102
-size 3913
