--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ExamplePreviewsScreenshots/ButtonPreviewTest_Dark_d19fbf1f_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:ddb53e3740e230c29513dfd1ee6f4bdb81e0f6966b81ad3be5694001aaab7280
-size 4300
