--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ExamplePreviewsScreenshots/ButtonPreviewTest_Landscape Mode_67531073_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:79224705ca2a6bee90bcf71352af4676737e82ecb99eaf79e84b47ede00d9837
-size 1999
