--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ExamplePreviewsScreenshots/ButtonPreviewTest_Large Font Size_8b5d39f1_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:03f813f86967978fcf9a93505aa09dcdfd12fa20a36b43b25d011db6e9c6f538
-size 5975
