--- app/src/debug/screenshotTest/reference/com/demo/utech/screenshottesting/ExamplePreviewsScreenshots/ButtonPreviewTest_Portrait Mode_3f51852e_0.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:521edd5dd50d6e622dbc45da3318e26e43d50ed0c557aff5d72745aee2ec84b9
-size 4417
