+++ app/src/main/java/com/demo/utech/screenshottesting/ui/components/CallCard.kt
@@ -1,4 +1,4 @@
-package com.demo.utech.screenshottesting.ui
+package com.demo.utech.screenshottesting.ui.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
