+++ app/src/main/java/com/demo/utech/screenshottesting/ui/model/Contact.kt
@@ -1,0 +1,6 @@
+package com.demo.utech.screenshottesting.ui.model
+
+data class Contact(
+    val name: String,
+    val phoneNumber: String,
+)
