+++ app/src/main/res/values-fr/strings.xml
@@ -6,5 +6,5 @@
     <string name="recent">Récent</string>
     <string name="directory">Annuaire</string>
     <string name="call_name">Appeler %1$s</string>
-    <string name="click_me">Click ici</string>
+    <string name="click_me">Cliquez moi!</string>
 </resources>
