+++ app/src/screenshotTest/kotlin/com/demo/utech/screenshottesting/DirectoryScreenScreenshots.kt
@@ -2,12 +2,12 @@ package com.demo.utech.screenshottesting
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.demo.utech.screenshottesting.ui.Contact
-import com.demo.utech.screenshottesting.ui.ContactProvider
+import com.demo.utech.screenshottesting.ui.data.ContactProvider
 import com.demo.utech.screenshottesting.ui.directory.DirectoryScreenPreview
+import com.demo.utech.screenshottesting.ui.model.Contact
 
 class DirectoryScreenScreenshots {
-    @PreviewAnnotations
+    @PreviewScreenshotTest
     @Composable
     fun CallCardPreviewTest(
         @PreviewParameter(ContactProvider::class) directory: List<Contact>
