+++ app/src/screenshotTest/kotlin/com/demo/utech/screenshottesting/PreviewScreenshotTest.kt
@@ -2,9 +2,9 @@ package com.demo.utech.screenshottesting
 
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import com.demo.utech.screenshottesting.ui.PreviewFrench
-import com.demo.utech.screenshottesting.ui.PreviewsUFontScale
-import com.demo.utech.screenshottesting.ui.PreviewsUOrientation
+import com.demo.utech.screenshottesting.ui.theme.PreviewFrench
+import com.demo.utech.screenshottesting.ui.theme.PreviewsUFontScale
+import com.demo.utech.screenshottesting.ui.theme.PreviewsUOrientation
 
 /**
  * Annotation for composing screenshot tests.
@@ -16,4 +16,4 @@ import com.demo.utech.screenshottesting.ui.PreviewsUOrientation
 @PreviewsUOrientation
 @PreviewsUFontScale
 @PreviewLightDark
-annotation class PreviewAnnotations
+annotation class PreviewScreenshotTest
